+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [0.13.0] - 2022-05-30
+### Changed
+* Update dependencies
+* Reject invalid query parameter "n"
+* Reject query parameter "n" > 1 for now (for feature detection)
+* Disable JavaScript in /noscript using Content-Security-Policy
+* Allow cross-origin request to private network (Private Network Access Preflight)
+
 ## [0.12.1] - 2022-03-12
 ### Changed
 * Update dependencies
@@ -217,7 +225,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ### Added
 * Implement basic Piping Server
 
-[Unreleased]: https://github.com/nwtgck/piping-server-rust/compare/v0.12.1...HEAD
+[Unreleased]: https://github.com/nwtgck/piping-server-rust/compare/v0.13.0...HEAD
+[0.13.0]: https://github.com/nwtgck/piping-server-rust/compare/v0.12.1...v0.13.0
 [0.12.1]: https://github.com/nwtgck/piping-server-rust/compare/v0.12.0...v0.12.1
 [0.12.0]: https://github.com/nwtgck/piping-server-rust/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/nwtgck/piping-server-rust/compare/v0.10.2...v0.11.0